Update list parsers to new method
