+++ src/data/lists/scoreboard-slot.ts
@@ -1,0 +1,14 @@
+import { COLORS } from "../../colors";
+
+export const scoreboardSlots = createSlots();
+
+function createSlots(): string[] {
+    const slots = [];
+
+    slots.push("list", "sidebar", "belowName");
+
+    for (const s of COLORS) {
+        slots.push(`sidebar.team.${s}`);
+    }
+    return slots;
+}
