+++ src/data/lists/statics.ts
@@ -1,0 +1,127 @@
+import { COLORS } from "../../colors";
+
+export const anchors = ["feet", "eyes"];
+export const operations = ["+=", "-=", "*=", "/=", "%=", "=", ">", "<", "><"];
+export const colors = COLORS;
+
+export const effects = [
+    "minecraft:speed",
+    "minecraft:slowness",
+    "minecraft:haste",
+    "minecraft:mining_fatigue",
+    "minecraft:strength",
+    "minecraft:instant_health",
+    "minecraft:instant_damage",
+    "minecraft:jump_boost",
+    "minecraft:nausea",
+    "minecraft:regeneration",
+    "minecraft:resistance",
+    "minecraft:fire_resistance",
+    "minecraft:water_breathing",
+    "minecraft:invisibility",
+    "minecraft:blindness",
+    "minecraft:night_vision",
+    "minecraft:hunger",
+    "minecraft:weakness",
+    "minecraft:poison",
+    "minecraft:wither",
+    "minecraft:health_boost",
+    "minecraft:absorption",
+    "minecraft:saturation",
+    "minecraft:glowing",
+    "minecraft:levitation",
+    "minecraft:luck",
+    "minecraft:unluck",
+    "minecraft:slow_falling",
+    "minecraft:conduit_power",
+    "minecraft:dolphins_grace"
+];
+
+export const enchantments = [
+    "minecraft:protection",
+    "minecraft:fire_protection",
+    "minecraft:feather_falling",
+    "minecraft:blast_protection",
+    "minecraft:projectile_protection",
+    "minecraft:respiration",
+    "minecraft:aqua_affinity",
+    "minecraft:thorns",
+    "minecraft:depth_strider",
+    "minecraft:frost_walker",
+    "minecraft:binding_curse",
+    "minecraft:sharpness",
+    "minecraft:smite",
+    "minecraft:bane_of_arthropods",
+    "minecraft:knockback",
+    "minecraft:fire_aspect",
+    "minecraft:looting",
+    "minecraft:sweeping",
+    "minecraft:efficiency",
+    "minecraft:silk_touch",
+    "minecraft:unbreaking",
+    "minecraft:fortune",
+    "minecraft:power",
+    "minecraft:punch",
+    "minecraft:flame",
+    "minecraft:infinity",
+    "minecraft:luck_of_the_sea",
+    "minecraft:lure",
+    "minecraft:loyalty",
+    "minecraft:impaling",
+    "minecraft:riptide",
+    "minecraft:channeling",
+    "minecraft:mending",
+    "minecraft:vanishing_curse"
+];
+
+export const particles = [
+    "minecraft:ambient_entity_effect",
+    "minecraft:angry_villager",
+    "minecraft:barrier",
+    "minecraft:block",
+    "minecraft:bubble",
+    "minecraft:cloud",
+    "minecraft:crit",
+    "minecraft:damage_indicator",
+    "minecraft:dragon_breath",
+    "minecraft:dripping_lava",
+    "minecraft:dripping_water",
+    "minecraft:dust",
+    "minecraft:effect",
+    "minecraft:elder_guardian",
+    "minecraft:enchanted_hit",
+    "minecraft:enchant",
+    "minecraft:end_rod",
+    "minecraft:entity_effect",
+    "minecraft:explosion_emitter",
+    "minecraft:explosion",
+    "minecraft:falling_dust",
+    "minecraft:firework",
+    "minecraft:fishing",
+    "minecraft:flame",
+    "minecraft:happy_villager",
+    "minecraft:heart",
+    "minecraft:instant_effect",
+    "minecraft:item",
+    "minecraft:item_slime",
+    "minecraft:item_snowball",
+    "minecraft:large_smoke",
+    "minecraft:lava",
+    "minecraft:mycelium",
+    "minecraft:note",
+    "minecraft:poof",
+    "minecraft:portal",
+    "minecraft:rain",
+    "minecraft:smoke",
+    "minecraft:spit",
+    "minecraft:squid_ink",
+    "minecraft:sweep_attack",
+    "minecraft:totem_of_undying",
+    "minecraft:underwater",
+    "minecraft:splash",
+    "minecraft:witch",
+    "minecraft:bubble_pop",
+    "minecraft:current_down",
+    "minecraft:bubble_column_up",
+    "minecraft:nautilus"
+];
