--- src/parsers/minecraft/list/item-enchantment.ts
@@ -1,9 +1,0 @@
-import { CommandErrorBuilder } from "../../../brigadier/errors";
-import { ListParser } from "./list";
-
-const exc = new CommandErrorBuilder(
-    "enchantment.unknown",
-    "Unknown enchantment: %s"
-);
-
-export const parser = new ListParser("minecraft:item_enchantment", exc);
