--- src/parsers/minecraft/list/item-slot.ts
@@ -1,6 +1,0 @@
-import { CommandErrorBuilder } from "../../../brigadier/errors";
-import { ListParser } from "./list";
-
-const exc = new CommandErrorBuilder("slot.unknown", "Unknown slot '%s'");
-
-export const parser = new ListParser("minecraft:item_slot", exc);
