--- src/parsers/minecraft/list/lists/color.ts
@@ -1,7 +1,0 @@
-import { COLORS } from "../../../../colors";
-import { ListSupplier } from "../lists";
-
-export const colorSupplier: ListSupplier = {
-    get: () => COLORS,
-    init: () => undefined
-};
