--- src/parsers/minecraft/list/lists/enchantment.ts
@@ -1,43 +1,0 @@
-import { ListSupplier } from "../lists";
-
-const enchs = [
-    "minecraft:protection",
-    "minecraft:fire_protection",
-    "minecraft:feather_falling",
-    "minecraft:blast_protection",
-    "minecraft:projectile_protection",
-    "minecraft:respiration",
-    "minecraft:aqua_affinity",
-    "minecraft:thorns",
-    "minecraft:depth_strider",
-    "minecraft:frost_walker",
-    "minecraft:binding_curse",
-    "minecraft:sharpness",
-    "minecraft:smite",
-    "minecraft:bane_of_arthropods",
-    "minecraft:knockback",
-    "minecraft:fire_aspect",
-    "minecraft:looting",
-    "minecraft:sweeping",
-    "minecraft:efficiency",
-    "minecraft:silk_touch",
-    "minecraft:unbreaking",
-    "minecraft:fortune",
-    "minecraft:power",
-    "minecraft:punch",
-    "minecraft:flame",
-    "minecraft:infinity",
-    "minecraft:luck_of_the_sea",
-    "minecraft:lure",
-    "minecraft:loyalty",
-    "minecraft:impaling",
-    "minecraft:riptide",
-    "minecraft:channeling",
-    "minecraft:mending",
-    "minecraft:vanishing_curse"
-];
-
-export const enchantmentSupplier: ListSupplier = {
-    get: () => enchs,
-    init: () => undefined
-};
