--- src/parsers/minecraft/list/lists/entity-anchor.ts
@@ -1,8 +1,0 @@
-import { ListSupplier } from "../lists";
-
-const anchors = ["feet", "eyes"];
-
-export const entityAnchorSupplier: ListSupplier = {
-    get: () => anchors,
-    init: () => undefined
-};
