--- src/parsers/minecraft/list/lists/operation.ts
@@ -1,8 +1,0 @@
-import { ListSupplier } from "../lists";
-
-const ops = ["+=", "-=", "*=", "/=", "%=", "=", ">", "<", "><"];
-
-export const operationSupplier: ListSupplier = {
-    get: () => ops,
-    init: () => undefined
-};
