+++ src/test/blanks.ts
@@ -1,6 +1,5 @@
 import { GlobalData } from "../data/types";
 import { PackLocationSegments } from "../misc-functions";
-import { Lists } from "../parsers/minecraft/list/lists";
 import { CommmandData } from "../types";
 import { ReturnAssertionInfo, TestParserInfo } from "./assertions";
 
@@ -29,6 +28,5 @@ export const emptyGlobal: GlobalData = {
     commands: { type: "root" },
     items: [],
     meta_info: { version: "" },
-    resources: {},
-    static_lists: new Lists()
+    resources: {}
 };
